Removed need for client.id and client.secret params. Removed need for org.url param. Added -sandbox param as boolean defaulted to false.
Changed auth for tooling api to use the same token that
was returned in the ConnectorConfig sessionId.  Pulling the
Instance URL for the org from the ConnectorConfig too.  The
sandbox param is to determine if the auth endpoint is:
    login.salesforce.com or
    test.salesforce.com

Made the ConnectorConfig and PartnerConnectorConfig static
singleton instances so that the same config object is used
across all requests.  It is populated on the first request
and reused afterwards so that the sessionId remains.